Adds a new field, 'settlement_school'
In order to identify settlement schools and, potentially, new settlement
schools in the future, this adds a field to the Story model that will be
blank by default, but carry a school code if it's a settlement school.

The first school code is 'edmc', and it's applied to all EDMC schools in
our database. This will also allow us to identify when to apply
different language to a page, since it appears we'll need different
content in some areas for settlement and non-settlement views.