In this commit, we have a working budget section. This includes pulling the BLS data from our API based on the program's BLS region, pulling monthly loan payment from the student debt calculator outputs, pulling salary from the program data, and setting the BLS expenses based on that salary.
An important coding change in this commit is the beginning of separating the
dispatcher 'get-model-values' into one file per model, such as 'get-expenses-
values'. The reason for this is simple - the models need to communicate (e.g.
expenses needs to know monthly loan cost from financials to calculate total
expenses), and this caused require() loops (which are bad). Since `get-model-
values` require()s all three models, any model trying to require() it would
cause an immediate self-require() (which is bad). If the models tried to
directly require each other, it's partly possible to avoid a require() loop...
for now. But it seemed to me that it was really a bigger problem that 'get-
model-values' was over-reaching. Separate dispatchers are, I feel, a much
better long-term solution.

The work of removing 'get-model-values' isn't finished in this commit, but
it's a priority for my next commit.