Add a manage.py command for updating colleges
Exposing the script for updating college data via the collegeScorecard
API will make it easier for Jenkins or other automation tools to update
school data. The `update_via_api` command will attempt to update every
school in our database based on the `latest_year` value stored as a
constant in our database.

An optional flag of `--school_id` will limit an update to just one
school, which is useful for testing, or for fixing an error found with a
single school.