Add logging and messaging to Notification model
We need to track and retry notification failures, so we're adding a log
entry to each Notification record to track its own transmission attempts.
We're also adding a "sent" boolean that is True
if a notification was successful. In this way we can regularly poll the
database to see if there are send failures and retry them or, if they're
getting old, collect them for troubleshooting and for notifying schools.