Adds management commands for notifications
This adds management command wrappers for the retry_notifications
and send_stale_notifications scripts, to allow automation of
notification monitoring in Jenkins.

This also adds a 'days' parameter to retry_notifications in case we want
to change the number of days back to search for unsent notifications.
The default remains 1 day.