We need to catch rates that are in the wrong form
Our CSV spec failed to explicitly say that rates should be in the form
"0.55" instead of "55" to represent 55%. We just asked for floats, which
allows 0.55 and 55.0. Instead of republishing our spec (which is being
ignored anyway), this PR catches any rates that are expressed as >1
numbers and converts them.

The PR also catches an encoding issue that cropped up in the first batch
of programs. EDMC's CSV was not UTF-8 encoded, and it included some
control characters that sneak into Excel files instead of hyphens.
Both code and collegedata.json are updated here.