Add functions for picking up new schools
Our data updating automations didn't include a way for new schools to
get added to our database. This introduces functions in the ipeds_update
script that will create new schools when we encounter school IDs that
are not in our database.

This also changes our Notification and Contact models to allow for
multiple email contacts at a given school or school system.