Add a private loan fee value to our constants
Following up on a [PR
comment](https://github.com/cfpb/college-costs/pull/182#issuecomment-226769796)
by Patrick in June, this gives us a future home for private loan fees,
if they ever become a thing. The current value is 0, because we're not
aware of any private loans that are charging fee points.