Adjust salary variable names for clarity
Our variable names evolved from a blending of legacy code and new data
sources, and in many cases became confusing. Salary is especially so.
Since we need to be clear about which source is being used for salary
figures -- program data or school data -- This renames the json salary
variables to 'programSalary' and 'schoolSalary'.