Add constant to track year when constants are updated
Our "About this tool" page has a reference to the academic year for
which rates apply. This delivers that year, based on when constants were
updated. It requires the person who updates constants to change this value.

This PR also clarifies the naming of year constants to better reflect
their usage. `latestYear` becomes `apiYear` and `latestSalaryYear`
becomes `salaryYear`