Add handling for test programs
This adds an "as_csv" model function to programs to make it easy to
create test programs for demo purposes. It also adjusts the load_program
script to handle an optional "test" column so the programs can be loaded
by a management command via file or s3.

We also tweak the `/offer/test/` option so that it displays *only* test
programs.