Add a management command to tag settlement schools
We need a way to tag a new group of settlement schools as they come in,
and this provides a way to do it via CSV, without having to deploy code.
It also affords a way to *untag* schools that had been marked as
settlement schools for demo purposes. Schools in this state have a
settlement_school value of `demo`, and by setting up a CSV with all the
demo schools and providing a blank tag value, we can de-tag all the demo
schools in one go.