Improve message posted when updating colleges
This only affects a message a developer would seen when updating the
colleges database. It notes an important constant that needs to be
checked -- Salary year. It's a constant stored in the database, and it
differs from the 'Latest year' constant, because salary data is always a
couple years behind the rest of the data.