Add handling for a new disclosure URL field `leng`
Previously program length, the standard intended length for a course of
study, was provided by the school in its program data. The new URL field
allows a school to override the standard program length with a
personalized length based on circumstances of an individual student,
such as a student transferring in with credits.

This commit also fixes a directional error in the intro, changes the
"Print offer" button to simply "Print" and cleans up some dead view
code.