support optional testing of other databases
This change adds support for using the `DATABASE_URL` environment
variable to specify Django's dev/test database, based on the
[dj-database-url](https://github.com/kennethreitz/dj-database-url)
package.

It lets you run tests against e.g. Postgres:

```sh
$ DATABASE_URL=postgres://user@localhost/db ./pytest.sh
```

It adds dj-database-url==0.4.2 to the dev/test requirements as that
matches the current version in cfgov-refresh.

It also excludes the Django settings files from coverage because
otherwise the new logic would reduce the coverage of this repo.