Make the `update_via_api` job to use Scorecard's new `latest` value.
The College Scorecard API now offers a `latest` parameter to fetch
the most recent data available, which is a big improvement over
the former guessing game.

This also allows us to ditch the former requirement of specifying
a new API_year value each year via the Django admin.

Changes in `/data_sources/` and `/fixtures/` are reference files
for data provenance.