Pass the offer data to the feedback form for recording
The feedback form had previously aquired the offer data indirectly from the referrer URL. This change passes that data directly to the form in the form of the query string so that it can be recorded along with the message. This should fix the current issues of the referrer being blank by explicitly providing the data needed.